requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.6
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
gunicorn==20.0.4
whitenoise==5.2.0
psycopg2-binary==2.8.6
rave-python==1.2.9
python-dotenv==0.15.0
drf-spectacular==0.18.2